+++ RELEASE_NOTES.md
@@ -68,6 +68,15 @@ v1.0.0-alpha.xx
   objects, respectively.
   [#914](https://github.com/OpenAssetIO/OpenAssetIO/issues/914)
 
+- Migrated the `ManagerInterface`/`Manager` `flushCaches` method to C++
+  with Python bindings.
+  [#994](https://github.com/OpenAssetIO/OpenAssetIO/issues/994)
+
+- Added a call to `flushCaches` after every `openassetio.test.manager`
+  (aka `apiComplianceSuite`) test case, giving the manager plugin a
+  chance to clean up between tests.
+  [#994](https://github.com/OpenAssetIO/OpenAssetIO/issues/994)
+
 v1.0.0-alpha.12
 ---------------